Signed commits Added tooltip to ply recording
Added native file dialog for folder/yaml selection.

Updated ktx submodule

Update Submodule

Update Submodule

Removed submodule [submodule name]

Removed ImGuiFileDialog external dependency
Signed commits

+++ include/Viewer/Scripts/Objects/MultiSenseCamera.cpp
@@ -210,7 +210,7 @@ void MultiSenseCamera::onUIUpdate(VkRender::GuiObjectHandles *uiHandle) {
             rateTableIndex = d.rateTableIndex;
             auto ratesInfo = renderData.crlCamera->getCameraInfo(0).imuSensorInfos.front();
             for (const auto &rate: ratesInfo.rates)
-                rates.emplace_back(rate.sampleRate);
+                rates.emplace_back(static_cast<uint32_t>(rate.sampleRate));
 
             snprintf(sampleRateLabel, sizeof(sampleRateLabel), "Rate (Hz): %d", rates[rateTableIndex]);
